Add req query verification.
